+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+* Add support for all asc/desc sorting order combinations, including null placement:
+  `:asc_nulls_first`, `:asc_nulls_last`, `:desc_nulls_first`, and `:desc_nulls_last`.
+  ([#136](https://github.com/duffelhq/paginator/pull/136))
+
 ## v1.0.4 - 2021-03-15
 
 * Fix type errors, thanks! @djthread: